+++ plugins/Organic/Organic.cpp
@@ -60,7 +60,7 @@ Plugin::Descriptor PLUGIN_EXPORT organic_plugin_descriptor =
 
 }
 
-float * OrganicInstrument::s_harmonics = nullptr;
+double * OrganicInstrument::s_harmonics = nullptr;
 
 /***********************************************************************
 *
@@ -103,36 +103,36 @@ OrganicInstrument::OrganicInstrument( InstrumentTrack * _instrument_track ) :
 
 	}
 
-/*	m_osc[0]->m_harmonic = log2f( 0.5f );	// one octave below
-	m_osc[1]->m_harmonic = log2f( 0.75f );	// a fifth below
-	m_osc[2]->m_harmonic = log2f( 1.0f );	// base freq
-	m_osc[3]->m_harmonic = log2f( 2.0f );	// first overtone
-	m_osc[4]->m_harmonic = log2f( 3.0f );	// second overtone
-	m_osc[5]->m_harmonic = log2f( 4.0f );	// .
-	m_osc[6]->m_harmonic = log2f( 5.0f );	// .
-	m_osc[7]->m_harmonic = log2f( 6.0f );	// .*/
+/*	m_osc[0]->m_harmonic = log2( 0.5 );	// one octave below
+	m_osc[1]->m_harmonic = log2( 0.75 );	// a fifth below
+	m_osc[2]->m_harmonic = log2( 1.0 );	// base freq
+	m_osc[3]->m_harmonic = log2( 2.0 );	// first overtone
+	m_osc[4]->m_harmonic = log2( 3.0 );	// second overtone
+	m_osc[5]->m_harmonic = log2( 4.0 );	// .
+	m_osc[6]->m_harmonic = log2( 5.0 );	// .
+	m_osc[7]->m_harmonic = log2( 6.0 );	// .*/
 
 	if( s_harmonics == nullptr )
 	{
-		s_harmonics = new float[ NUM_HARMONICS ];
-		s_harmonics[0] = log2f( 0.5f );
-		s_harmonics[1] = log2f( 0.75f );
-		s_harmonics[2] = log2f( 1.0f );
-		s_harmonics[3] = log2f( 2.0f );
-		s_harmonics[4] = log2f( 3.0f );
-		s_harmonics[5] = log2f( 4.0f );
-		s_harmonics[6] = log2f( 5.0f );
-		s_harmonics[7] = log2f( 6.0f );
-		s_harmonics[8] = log2f( 7.0f );
-		s_harmonics[9] = log2f( 8.0f );
-		s_harmonics[10] = log2f( 9.0f );
-		s_harmonics[11] = log2f( 10.0f );
-		s_harmonics[12] = log2f( 11.0f );
-		s_harmonics[13] = log2f( 12.0f );
-		s_harmonics[14] = log2f( 13.0f );
-		s_harmonics[15] = log2f( 14.0f );
-		s_harmonics[16] = log2f( 15.0f );
-		s_harmonics[17] = log2f( 16.0f );
+		s_harmonics = new double[ NUM_HARMONICS ];
+		s_harmonics[0] = log2( 0.5 );
+		s_harmonics[1] = log2( 0.75 );
+		s_harmonics[2] = log2( 1.0 );
+		s_harmonics[3] = log2( 2.0 );
+		s_harmonics[4] = log2( 3.0 );
+		s_harmonics[5] = log2( 4.0 );
+		s_harmonics[6] = log2( 5.0 );
+		s_harmonics[7] = log2( 6.0 );
+		s_harmonics[8] = log2( 7.0 );
+		s_harmonics[9] = log2( 8.0 );
+		s_harmonics[10] = log2( 9.0 );
+		s_harmonics[11] = log2( 10.0 );
+		s_harmonics[12] = log2( 11.0 );
+		s_harmonics[13] = log2( 12.0 );
+		s_harmonics[14] = log2( 13.0 );
+		s_harmonics[15] = log2( 14.0 );
+		s_harmonics[16] = log2( 15.0 );
+		s_harmonics[17] = log2( 16.0 );
 	}
 
 	for (int i=0; i < m_numOscillators; i++) {
@@ -603,10 +603,10 @@ void OscillatorObject::updateVolume()
 
 void OscillatorObject::updateDetuning()
 {
-	m_detuningLeft = powf( 2.0f, OrganicInstrument::s_harmonics[ static_cast<int>( m_harmModel.value() ) ]
+	m_detuningLeft = pow( 2.0, OrganicInstrument::s_harmonics[ static_cast<int>( m_harmModel.value() ) ]
 				+ (float)m_detuneModel.value() * CENT ) /
 				Engine::audioEngine()->processingSampleRate();
-	m_detuningRight = powf( 2.0f, OrganicInstrument::s_harmonics[ static_cast<int>( m_harmModel.value() ) ]
+	m_detuningRight = pow( 2.0, OrganicInstrument::s_harmonics[ static_cast<int>( m_harmModel.value() ) ]
 				- (float)m_detuneModel.value() * CENT ) /
 				Engine::audioEngine()->processingSampleRate();
 }
